backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Vector Search
sqlite-vss>=0.1.2
numpy>=1.26.3

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File Handling
python-multipart>=0.0.6
aiofiles>=23.2.1
python-magic>=0.4.27

# Document Parsing
beautifulsoup4>=4.12.3
markdown>=3.5.2
pypdf>=3.17.4
Pillow>=10.2.0
pytesseract>=0.3.10
ebooklib>=0.18
lxml>=5.1.0

# HTTP Client (for HuggingFace API)
httpx>=0.26.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
black>=23.12.1
ruff>=0.1.11
mypy>=1.8.0
